+++ CHANGELOG.md
@@ -5,8 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## Unreleased
+
+
+### Changed
+
+- Rich will display tracebacks with finely grained error locations on python 3.11+ https://github.com/Textualize/rich/pull/3486
+
+
+### Fixed
+
+- Fixed issue with Segment._split_cells https://github.com/Textualize/rich/pull/3506
+
+### Added
+
+- Add a new `column` object `IterationSpeedColumn`. https://github.com/Textualize/rich/pull/3332
+
+## [13.8.1] - 2024-09-10
+
+### Fixed
+
+- Added support for Python 3.13 https://github.com/Textualize/rich/pull/3481
+- Fixed infinite loop when appending Text to same instance https://github.com/Textualize/rich/pull/3480
+
+## [13.8.0] - 2024-08-26
 
 ### Fixed
 
@@ -17,6 +39,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Progress track thread is now a daemon thread https://github.com/Textualize/rich/pull/3402
 - Fixed cached hash preservation upon clearing meta and links https://github.com/Textualize/rich/issues/2942
 - Fixed overriding the `background_color` of `Syntax` not including padding https://github.com/Textualize/rich/issues/3295
+- Fixed pretty printing of dataclasses with a default repr in Python 3.13 https://github.com/Textualize/rich/pull/3455
+- Fixed selective enabling of highlighting when disabled in the `Console` https://github.com/Textualize/rich/issues/3419
+- Fixed BrokenPipeError writing an error message https://github.com/Textualize/rich/pull/3468
+- Fixed superfluous space above Markdown tables https://github.com/Textualize/rich/pull/3469
+- Fixed issue with record and capture interaction https://github.com/Textualize/rich/pull/3470
+- Fixed control codes breaking in `append_tokens` https://github.com/Textualize/rich/pull/3471
+- Fixed exception pretty printing a dataclass with missing fields https://github.com/Textualize/rich/pull/3472
 
 ### Changed
 
@@ -30,7 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Adds a `case_sensitive` parameter to `prompt.Prompt`. This determines if the
   response is treated as case-sensitive. Defaults to `True`.
-- Add a new `column` object `IterationSpeedColumn`. https://github.com/Textualize/rich/pull/3332
+- Added `Console.on_broken_pipe` https://github.com/Textualize/rich/pull/3468
 
 ## [13.7.1] - 2024-02-28
 
@@ -2053,6 +2082,8 @@ Major version bump for a breaking change to `Text.stylize signature`, which corr
 
 - First official release, API still to be stabilized
 
+[13.8.1]: https://github.com/textualize/rich/compare/v13.8.0...v13.8.1
+[13.8.0]: https://github.com/textualize/rich/compare/v13.7.1...v13.8.0
 [13.7.1]: https://github.com/textualize/rich/compare/v13.7.0...v13.7.1
 [13.7.0]: https://github.com/textualize/rich/compare/v13.6.0...v13.7.0
 [13.6.0]: https://github.com/textualize/rich/compare/v13.5.3...v13.6.0